Move shell to the kernel (sorta hardcoded)
